+++ components/mongodb/package.json
@@ -15,6 +15,6 @@
         "access": "public"
     },
     "dependencies": {
-        "mongoose": "^6.1.6"
+        "mongodb": "^4.6.0"
     }
 }
