+++ components/mongodb/sources/new-field-in-document/new-field-in-document.mjs
@@ -1,10 +1,10 @@
-import common from "../common.mjs";
+import common from "../common/base.mjs";
 
 export default {
   ...common,
   key: "mongodb-new-field-in-document",
   name: "New Field in Document",
-  description: "Emits an event when a new field is added to a document",
+  description: "Emit new an event when a new field is added to a document",
   version: "0.0.2",
   type: "source",
   dedupe: "unique",
@@ -48,7 +48,7 @@ export default {
       return !fields.includes(key);
     },
     async processEvent(client, ts) {
-      let fields = this._getFields() || [];
+      const fields = this._getFields() || [];
       const collection = this.mongodb.getCollection(client, this.database, this.collection);
       const documents = await this.mongodb.listDocuments(collection);
       const doc = documents.find((obj) => {
@@ -56,7 +56,9 @@ export default {
       });
       const keys = Object.keys(doc);
       for (const key of keys) {
-        if (!this.isRelevant(key, fields)) continue;
+        if (!this.isRelevant(key, fields)) {
+          continue;
+        }
         fields.push(key);
         this.emitEvent(key, ts);
       }
